requirements.txt
anthropic==0.49.0
python-telegram-bot==21.10
httpx==0.28.1
redis==5.2.1
pydantic-settings==2.8.2
structlog==25.1.0
tenacity==9.0.0
python-dotenv==1.0.1
